Only synthesize grapheme widths once for surrogate pair characters
This patch updates the grapheme width synthesis code to not set
surrogate pairs to be half-width. With this change our rendering
more closely matches Gecko.

There are no pixel differences in svg/text layouttests as a result
of this codechange. This patch does unify the multiple copies of
highlight code which resulted in a small pixel difference in
combining-character-queries.html due to the height of the highlight
text.

BUG=597311

Review URL: https://codereview.chromium.org/1847763002

Cr-Original-Commit-Position: refs/heads/master@{#384412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec8e7dc866c0edc91cb9c79cd82ea40d2b5e5b58