Run auto-formatter (autopep8) on webkitpy.
This is a follow-up to https://codereview.chromium.org/1783073002 (run autopep8 on webkitpy/layout_tests).

The main reason for this is that it increases code consistency.

It would also make it easier to turn on pylint warnings when running check-webkit-style and the presubmit (http://crbug.com/598897),
because it decreases the number of warnings related to formatting and whitespace.

Review URL: https://codereview.chromium.org/1839193004

Cr-Original-Commit-Position: refs/heads/master@{#384428}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4794e48d39b9d037ed247c73135631100189d541