Make sure to save/restore scroll positions in finishDelayUpdateScrollInfo
Due to Flexbox's multi-pass layout, we may lay out children at
intermediate sizes that are very small, causing us to clamp the
scroll position at this small size. Once we lay out at the right
size, the scroll position is lost.

To avoid this problem, make finishDelayUpdateScrollInfo store
the old scroll position so it can be restored.

XXXXX THIS IS TEMPORARY XXXXX

The plan is to bake this on canary and merge this to M50 to
fix the release blocker and fix this the right way on trunk
(bug 600036), by delaying clamping of scroll offsets to after layout.

BUG=593209

Review URL: https://codereview.chromium.org/1846023003

Cr-Original-Commit-Position: refs/heads/master@{#384772}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea799d37bd27d7ae346031a9332f542f7f57fd38