Fix indentation in template for [OriginTrialEnabled]
This is the first CL in a series to cleanup the bindings tests/results
for the [OriginTrialEnabled] attribute. The cleanups are in separate CLs
to simplify the final CL to make the [OriginTrialEnabled] and
[RuntimeEnabled] attributes mutually exclusive.

This CL just corrects the indentation in generated code for the
[OriginTrialEnabled] attribute. The template had extra indentation
which caused a mismatch with surrounding code.

CLs in the series:
1) This CL, Fix indentation in template
2) https://codereview.chromium.org/1830113003/, Rename test IDL members
3) https://codereview.chromium.org/1838813002/, Separate bindings tests for [OriginTrialEnabled]

BUG=586594

Review URL: https://codereview.chromium.org/1828973003

Cr-Original-Commit-Position: refs/heads/master@{#385160}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 669cb1aa76222f4ea77a88d842c4d52814735b19