Track whether an element was inserted via document.write
This patch adds a member to ScriptLoader which tracks whether elements are created
via a call to document.write. These scripts are going to be parser blocking. The plan is to use this data for metrics tracking, as well as driving a prediction
service which will learn to preload these scripts when we load their initiator.

BUG=594159

Review URL: https://codereview.chromium.org/1814853003

Cr-Original-Commit-Position: refs/heads/master@{#385218}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93bafc396d0c61800ad30d8dc569539ca3de7bd9