Crash on underflow in ImageResource.cpp + fixed where it was triggered.
Previously there was a bug where BitmapImage was not tracking its memory
usage correctly and was underflowing a value. When I added a check to
ensure it wasn't underflowing it uncovered a crash with animated gifs.
This patch fixes that underflow as well.

Review URL: https://codereview.chromium.org/1838633002

Cr-Original-Commit-Position: refs/heads/master@{#385227}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 03d2937ebf3445d183c8c325082efee772360e9d