[css-grid] Remove unnecessary iteration in populateGridPositions loop.
The populateGridPositions loop limit was set to 'lastLine'. However, the
the position of last track's start line is updated after the loop, since
it does not follow the same pattern; it does not have a content
distribution offset.

So, since we are essentially overwriting the value stored in the last
iteration, we can just lower the loop limit.

No new tests added, because there is no change in the functionality.

Review URL: https://codereview.chromium.org/1858193004

Cr-Original-Commit-Position: refs/heads/master@{#385700}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f1cd3a47c2027e5c6bb66efd15d7239ed1af0fb