Fetch: Fix 'body' processing in RequestInit.
https://codereview.chromium.org/1844053003 broke 'RequestInit's ability
to process a body on a request with 'credentials' set because it was
bypassing 'body' processing if any 'credentials' property was set.

It _should_ have been bypassing 'body' processing if and only if the
'credentials' attribute contained a 'PasswordCredential' object.

This patch fixes the broken comparison.

BUG=599597

Review URL: https://codereview.chromium.org/1862293003

Cr-Original-Commit-Position: refs/heads/master@{#385716}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93a08a4b973b9cca6f7f3d182d9a13102755da2b