[css-grid] Fix definite/indefinite size detection for abspos elements
We were considering that any abspos element has a definite size, and
that's not true. That's only true if the offset properties in that
dimension (left and right or top and bottom) are non-auto.
Fixed LayoutBox::hasDefiniteLogicalWidth|Height() to check this
properly.

This has been confirmed by the CSS WG in the following thread:
https://lists.w3.org/Archives/Public/www-style/2015Nov/0074.html

BUG=538513
TEST=fast/css-grid-layout/positioned-grid-container-percentage-tracks.html

Review URL: https://codereview.chromium.org/1383003002

Cr-Original-Commit-Position: refs/heads/master@{#386045}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b75b9bdaa3ede6fe0230ae822885d6fb38ed514b