Make structured clone work with ImageBitmapOptions
At this moment, structured clone an ImageBitmap works fine with an ImageBitmap
created with the default options. Now that we have options of flipY and
premultiplyAlpha, we should make sure that the transfer logic works.

To verify that, we create an ImageBitmap with flipY=true and premultiplyAlpha
=false, we send this ImageBitmap to the worker thread via postMessage and
worker send it back to the main thread. Now on the main thread, the
ImageBitmap should have the same pixel data as the originally created one.

A layout test is added to make sure that the ImageBitmap's pixel data
survives the round trip from main<-->worker thread.

Review URL: https://codereview.chromium.org/1855803002

Cr-Original-Commit-Position: refs/heads/master@{#386233}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2128130250de3702dd0e8bcc3c09396e769b3b1f