generateReferrer shouldn't return a referrer whose policy is default
As discussed in [1], this CL changes SecurityPolicy::generateReferrer
so that the returned referrer's referrer policy will never be default.

We have reduce-referrer-granularity flag, which changes the default referrer
policy behavior so that it omits referrer path information when requesting
to a cross-origin site *in addition to* the usual behavior (which is equal
to ReferrerPolicyNoReferrerWhenDowngrade). This CL introduces a new referrer
policy ReferrerPolicyNoReferrerWhenDowngradeOriginOnlyForCrossOrigin
for the behavior.

1: https://github.com/whatwg/fetch/issues/266
BUG=532385

Review URL: https://codereview.chromium.org/1850083002

Cr-Original-Commit-Position: refs/heads/master@{#387039}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b5788fb4d238b4d2e26f8059d3d1485e278b8da