Convert main-thread overscroll parameters into viewport space.
The overscroll functions in the Browser expect these values to be in the
viewport coordinate space. i.e. If the page is scaled by a factor of 2,
if the user moves their finger by 10px and all of it is consumed by overscroll,
the overscroll method should get back a delta of 10px, even though we tried
scrolling the root frame by 5px.

More generally, coordinates coming out of Blink should almost always be in
viewport/renderer/widget coordinates since it's most useful in the renderer and
it's difficult for client code to compensate for the visual viewport transform
since that's an internal implementation detail of Blink (mostly).

Also cleaned up the naming of these parameters through the call chain.

BUG=602476

Review URL: https://codereview.chromium.org/1879103002

Cr-Original-Commit-Position: refs/heads/master@{#387154}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e53a10f9b029bf4367567a170fc77d6d6d42fe91