Oilpan: Remove ENABLE(OILPAN) from core/html/parser/
This CL adds a pre-finalizer to HTMLDocumentParser to unconditionally call
HTMLDocumentParser::detach(). Before this CL, HTMLDocumentParser::detach()
was not called when the document and the parser died at the same time.
This wouldn't be a right behavior since PendingScript::stopWatchingForLoad
needs to be unconditionally called (for example).

BUG=585328

Review URL: https://codereview.chromium.org/1883743002

Cr-Original-Commit-Position: refs/heads/master@{#387241}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 322e44b7527d8211d19558b5f2582e6b1a327ae4